fix fib return type

+++ examples/test.c
@@ -13,11 +13,11 @@ i32  eth_getCallDataSize() __attribute__((import_module("ethereum"),import_name(
 void eth_callDataCopy(void *res, i32 dOff, i32 dLen) __attribute__((import_module("ethereum"),import_name("callDataCopy")));
 */
 
-static u32 fib(u32 n) {
+static u64 fib(u32 n) {
 	if (n < 2) return n;
-	u32	result=2;
-	u32	pre = 1;
-	u32 next = 1;
+	u64	result=2;
+	u64	pre = 1;
+	u64	next = 1;
 	for (u32 i = 2; i < n; ++i) {
 		result = pre + next;
 		pre = next;
